# enable Half for cat serial kernel